Fix 'dummy' HPSS ioctl code
In order for the unit test suite to run happily without HPSS installed,
some dummy implementation of the ioctl() call normally given by the
'hpssfs' module needed to be added. This was done when bringing the code
into a known good state, but the implementation would not actually work
and that seemed like it would be an issue down the road. So here's an
actually working implementation of that.

Change-Id: I52c8ede8ac2d38bb59c2d7d2fb399fd65e75577a